requirements.txt
Django==1.8.5
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
defusedxml==0.4.1
dj-database-url==0.4.0
django-braces==1.8.1
django-dotenv==1.3.0
django-jinja==1.4.1
django-jinja-bootstrap-form==4.1.1
gunicorn==19.3.0
oauthlib==1.0.3
psycopg2==2.6.1
python-social-auth==0.2.13
python3-openid==3.0.7
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
selenium==2.48.0
six==1.10.0
whitenoise==2.0.6
